Adapt the build for 0.11 upon cherry pick
