+++ api/src/test/java/org/apache/iceberg/AssertHelpers.java
@@ -20,6 +20,9 @@
 package org.apache.iceberg;
 
 import java.util.concurrent.Callable;
+
+import org.assertj.core.api.AbstractThrowableAssert;
+import org.assertj.core.api.Assertions;
 import org.junit.Assert;
 
 public class AssertHelpers {
@@ -135,6 +138,26 @@ public class AssertHelpers {
     } catch (AssertionError e) {
       e.addSuppressed(actual);
       throw e;
+    }
+  }
+
+  public static void assertThrowsWithCause(
+          String message,
+          Class<? extends Exception> expected,
+          String expectedContainedInMessage,
+          Class<? extends Exception> cause,
+          String causeContainedInMessage,
+          Runnable runnable) {
+    AbstractThrowableAssert<?, ?> chain =
+            Assertions.assertThatThrownBy(runnable::run).as(message).isInstanceOf(expected);
+
+    if (expectedContainedInMessage != null) {
+      chain = chain.hasMessageContaining(expectedContainedInMessage);
+    }
+
+    chain = chain.getCause().isInstanceOf(cause);
+    if (causeContainedInMessage != null) {
+      chain.hasMessageContaining(causeContainedInMessage);
     }
   }
 }
