+++ build.gradle
@@ -129,6 +129,7 @@ subprojects {
     testCompile 'org.slf4j:slf4j-simple'
 
     testCompile 'org.mockito:mockito-core'
+    testCompile 'org.assertj:assertj-core'
   }
 
   test {
