+++ versions.props
@@ -23,10 +23,11 @@ javax.ws.rs:javax.ws.rs-api = 2.1.1
 io.quarkus:* = 1.9.1.Final
 
 # test deps
-junit:junit = 4.12
+junit:junit = 4.13.2
 org.mockito:mockito-core = 1.10.19
 org.apache.hive:hive-exec = 2.3.7
 org.apache.hive:hive-service = 2.3.7
 org.apache.tez:tez-dag = 0.8.4
 org.apache.tez:tez-mapreduce = 0.8.4
 com.adobe.testing:s3mock-junit4 = 2.1.28
+org.assertj:assertj-core = 3.23.1
